requirements.txt
flask
Flask-SocketIO
librosa
numpy
pydub
PyYAML
sounddevice
openai-whisper
python-dotenv
